UPSTREAM: mwifiex: set passive scan type for scan requests with no ssid
It was observed that station would sent probe request even when
scan type has been set as passive during iw scan.
This was happening because driver sets passive scan type only
when channel has IEEE80211_CHAN_PASSIVE_SCAN flag set.
Along with this, add condition to check if no ssids are specified in
scan request so as to mark such scan request passive.

Bug=none
Test=Set passive parameter while scanning using iw and ensure probe
request is not sent on specified channel.

Change-Id: I898bf93511cc266bc5b03fe5d3fb1dc731041e36
Signed-off-by: Avinash Patil <patila@marvell.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>
Reviewed-on: https://chromium-review.googlesource.com/216622
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256310