UPSTREAM: ALSA: seq: bind seq driver automatically
Currently the sequencer module binding is performed independently from
the card module itself.  The reason behind it is to keep the sequencer
stuff optional and allow the system running without it (e.g. for using
PCM or rawmidi only).  This works in most cases, but a remaining
problem is that the binding isn't done automatically when a new driver
module is probed.  Typically this becomes visible when a hotplug
driver like usb audio is used.

This patch tries to address this and other potential issues.  First,
the seq-binder (seq_device.c) tries to load a missing driver module at
creating a new device object.  This is done asynchronously in a workq
for avoiding the deadlock (modprobe call in module init path).

This action, however, should be enabled only when the sequencer stuff
was already initialized, i.e. snd-seq module was already loaded.  For
that, a new function, snd_seq_autoload_init() is introduced here; this
clears the blocking of autoloading, and also tries to load all pending
driver modules.

Reported-by: Adam Goode <agoode@chromium.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(cherry picked from commit 68ab61084de3220e2fb0a698c890ba91decddc85)

BUG=chromium:499817
TEST=Boot with no MIDI devices attached. Go to http://webaudiodemos.appspot.com/midi-synth/index.html and connect a device. Verify that it appears in the MIDI_IN menu on the page.

Change-Id: Ia2d44461cab9682332053d71f96f03c75ce3ccbf
Reviewed-on: https://chromium-review.googlesource.com/277491
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Adam Goode <agoode@chromium.org>
Tested-by: Adam Goode <agoode@chromium.org>