BACKPORT: staging: android: binder: move to the "real" part of the kernel
Backport note: This is binder from Android's 3.10 kernel with some
fixups for 3.8

The Android binder code has been "stable" for many years now.  No matter
what comes in the future, we are going to have to support this API, so
might as well move it to the "real" part of the kernel as there's no
real work that needs to be done to the existing code.

BUG=brillo:353
TEST=booted on panther and tested with servicemanager

(cherry-picked from upstream Linus 777783e0abae3cab7555bb182776f9ffaa35631a)
Signed-off-by: Lee Campbell <leecam@chromium.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Conflicts:
	drivers/Kconfig
	drivers/Makefile
	drivers/android/binder.h
	drivers/staging/android/Kconfig
	drivers/staging/android/Makefile
	include/uapi/linux/android/binder.h

Change-Id: Ib0be48397291b18dad2711314e8efa1d3c67f5d2
Reviewed-on: https://chromium-review.googlesource.com/253142
Reviewed-by: Will Drewry <wad@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Lee Campbell <leecam@chromium.org>
Tested-by: Lee Campbell <leecam@chromium.org>