daisy_max98095: Enable DMIC in all playback cases
On max98090, DMIC enable register can not be safely toggled when
shutdown bit is 1. This causes a short audio drop in playback whenever a
capture stream starts. To overcome this problem, turn on DMIC enable
bits during all audio playback cases.
Captured audio is not sent outside of codec because SDOEN is not enabled.

BUG=chromium:415156
TEST=On daisy 2.0, peach_pit (using max98090). Play youtube in the
background. Check there is no audio drop when capture stream starts.
TEST=On peach_pi (using max98091). Do the same test as above. Also,
check dmic is not enabled (0x13 register is 0x20) when there is only
playback stream.

Change-Id: I9b17eacb071a6ee950b995761ad0091e46c917d3
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/232356