CHROMIUM: ALSA: ASoC: max98090: Enable DMIC at playback when selected.
NOTE: This patch should not be sent to upstream.

DMIC enable register can not be safely toggled when shutdown bit is 1.
This causes a short audio drop in playback whenever a capture stream
starts.
To overcome this problem, this patch adds dapm routes to turn on DMIC
enable bits during all audio playback cases when DMIC is selected.
Captured audio is not sent outside of codec because SDOEN is not
enabled.
When external mic is selected, remove those dapm routes.

BUG=chromium:415156
TEST=On daisy and pit, play youtube in the background.
Check there is no audio drop when capture stream starts.
Checks recording from DMIC and external mic both work.

Change-Id: If434a4ba73a64f46caa35ecb0810e78c4bc082f5
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/238674