CHROMIUM: iwl7000: mvm: fix RX with both chains
A previous commit broke RX with 2 chains for diversity. This had
an impact on throughput where we're using only a single stream
(11a/b/g APs, single stream APs, static SMPS).

In addition, this caused some issues with chain B and resulted in
RX deafness  in certain scenarios with the Core7 firmware.

BUG=chrome-os-partner:34821
TEST=test scenario with chain B transition

Change-Id: Ie7dc38b04b425fb722d3055ca784645a73e3cbb8
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/246473
Reviewed-by: Paul Stewart <pstew@chromium.org>