CHROMIUM: iwl7000/mac80211: let flush() drop packets when possible
mac80211: let flush() drop packets when possible [BUGFIX]

When roaming / suspending, it makes no sense to wait until
the transmit queues of the device are empty. In extreme
condition they can be starved (VO saturating the air), but
even in regular cases, it is pointless to delay the roaming
because the low level driver is trying to send packets to
an AP which is far away. We'd rather drop these packets and
let TCP retransmit if needed. This will allow to speed up
the roaming.

For suspend, the explanation is even more trivial.

BUG=chrome-os-partner:34666
TEST=try to reproduce the scenario in the bug

Change-Id: I3d2c64d647896489cb588f3dc7f519cd8b7418bf
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/239513
Reviewed-by: Paul Stewart <pstew@chromium.org>