CHROMIUM: cros_ec: Properly return error when get bad ec response
Currently when ec get response with error result, we just print
debug message and continue to process it as if it has the success
result. Usually the response will fail checksum test and casue
kernel error spam but if this is not the case, it might cause
some more serious problems.

This patch will make ec driver return EBADMSG when get an ec
response with error result.

BUG=chromium:444431
TEST=Run 'ectool version' in spring. No bad checksum message

Change-Id: Ic40f042a37b1ad533bac066842df15a2719d6500
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/240272
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>