CHROMIUM: pstore/ram: verify ramoops header before saving record
On some devices the persistent memory contains junk after a cold boot,
and /dev/pstore/dmesg-ramoops-* are created with random data which is
not the result of a kernel crash.

This patch adds a ramoops header check and skips any
persistent_ram_zone that does not have a valid header.

BUG=chromium:443764
TEST=platform_KernelErrorPaths

Change-Id: I045ad463fa699358ccc10b9c60757fb68b5b3394
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/242335
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Commit-Queue: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>