CHROMIUM: iwl7000: mac80211: handle drv_add_interface failures properly during reconfig
If any interface fails to be added to the driver in during reconfig,
we should remove all the successfully added interfaces and report
reconfig failure, so things can be cleaned up properly.  Failing to do
so can lead to subsequent failures and leave the drivers in a messed
up state.

BUG=chrome-os-partner:35485
TEST=cause reconfig to fail, for example by sending MAC_CONTEXT_CMD
     in iwlmvm with color 255

Change-Id: Iaed44a80c039c09d9ab7af54942ee293e513f152
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/246522
Reviewed-by: Paul Stewart <pstew@chromium.org>