CHROMIUM: drm/exynos: signal display fence when finished displaying buffer
as opposed to when swapping to the buffer. The fences need to be double
buffered for the synchronization to work like KDS worked before and to
protect fb being displayed from GPU rendering. This is the standand
of other DRM drivers and while right now userspace will not schedule
rendering too early because it uses page flip event, it will become
crucial if we possibly re-implement "fake flip" events.

BUG=none
TEST=build and boot daisy, run WebGL Aquarium

Change-Id: I573d8fe998fa203004683e0d8860d669b5535604
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/253932
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>