CHROMIUM: [media] s5p-mfc: Set current context in single place
Scattering the assignment over all the code executing hardware commands
is not justified in any way, makes it more difficult to read the code
and easier to make mistakes. In fact there is already one code path
which leads to driver loop trying to run the same context again and
again if issuing command fails.

This patch fixes the problem by setting current context as soon as it is
selected for next hardware run.

BUG=chromium:495960
TEST=youtube;apprtc

Change-Id: If4873a700e58f9e167fb164adf95dec927ec1997
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256900
Reviewed-by: Pawel Osciak <posciak@chromium.org>