CHROMIUM: [media] s5p-mfc: Use pointers to refer to contexts
This patch makes the driver use pointers instead of some artificial
indices into a static array in driver data to refer to contexts. This
has the benefit of code simplification and allows to remove the
hardcoded instance limit in future. In addition, it removes a potential
race point of updating the arrays of pointers with context pointers on
open and NULL on release.

BUG=chromium:495960
TEST=youtube;apprtc

Change-Id: I12a0bd414b92e3abac3355747ff63661db4f9c1a
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256908
Reviewed-by: Pawel Osciak <posciak@chromium.org>