CHROMIUM: [media] s5p-mfc: Keep first frame of VP8 for decoding
Currently the driver releases first buffer with VP8 stream after the
hardware parses the header (SEQ_DONE) and requires the userspace to
requeue the same buffer again for proper decoding. This is inconsistent
with existing behavior for H264 and V4L2 Codec API and causes
compatibility issues with other V4L2 codec drivers.

This patch makes the behavior consistent with H264, V4L2 Codec API and
other drivers by holding the buffer used for header parsing until it is
decoded if source codec is VP8.

BUG=chrome-os-partner:36701
TEST=youtube;shaka player;apprtc

Change-Id: I81dc1a62eb3eca4f90fd6974f88f794fa2277bee
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256762
Reviewed-by: Pawel Osciak <posciak@chromium.org>