CHROMIUM: [media] s5p-mfc: Send event after resolution change is processed
The resolution change event should be sent to userspace only after new
header have been parsed and new format is available to be fetched using
G_FMT ioctl. This patch modifies the driver to follow this convention
and moves the code signalling the event to SEQ_DONE event handler. While
at it, the CAPTURE queue state setting is also moved there, so that it
is happenning in a synchronized manner under dev->irqlock.

BUG=chromium:495960
TEST=youtube;apprtc

Change-Id: I7e813c4527fa798667dbb84d601aae7221a297d2
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/256909
Reviewed-by: Pawel Osciak <posciak@chromium.org>