CHROMIUM: [media] s5p-mfc: Try run in resume
Currently the driver does not try running scheduled contexts after
resume automatically, so the hardware will be stuck until the userspace
queues more work. However the correct handling of suspend/resume
requires the hardware to be restored to the same state as before
suspend, so if further runs are available the hardware should continue
processing them.

BUG=chromium:495960
TEST=youtube;apprtc

Change-Id: Ib1883f4e83635aeb5d76580ba6aa09dc56372931
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/282452
Reviewed-by: Pawel Osciak <posciak@chromium.org>