CHROMIUM: s5p-mfc: Copy timestamp and timecode in encoder.
Adding support to copy the timestamp and timecode of the
source buffers of the mfc encoder into its destination
buffers.

BUG=chrome-os-partner:30604
TEST= Tested using mainline application v4l2-mfc-encoder.
Altered the test application at the v4l_enq_buf call to add
the timestamp for the source buffer. Added prints to check
the timestamp of source buffer and destination buffer and
confirmed with values obtained, copying  was successful.

Change-Id: I0d44d6b7b867ce0adee213591185051ccfd715fc
Signed-off-by: Smitha T Murthy <smitha.t@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/261880
Reviewed-by: Haixia Shi <hshi@chromium.org>
Commit-Queue: Haixia Shi <hshi@chromium.org>
Tested-by: Haixia Shi <hshi@chromium.org>