CHROMIUM: tpm: record TPM resets between request and response
NOT FOR UPSTREAMING - NOT FOR REBASING - KERNEL 3.8 ONLY (snow/lucas)

This sets a bit in the non-volatile vboot context while a TPM command
is executing, i.e. between request and response, in the first 30s after
boot.

The purpose is to be able to tell how many of the TPM DA counter
increases are due to such occurrences.  This is only of interest
for daisy/snow/lucas.

This bit is preserved across reboots and will be left alone if
set before the beginning of a command.

The functionality is accessed by registering callbacks to get
and set NVRAM bits from a platform-specific driver.

TESTING.  I used the instructions in 462428 #39 to produce a TPM reset
during a command, and observed (with crossystem) that the bit went
from 0 to 1 at the next reboot.

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

BUG=chromium:431360
TEST=see testing notes above.  Pretty hard to test.

Change-Id: I0cd44646a00a138b635e2983288073e792f11fed
Reviewed-on: https://chromium-review.googlesource.com/262465
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>