CHROMIUM: scan_swap_map: don't look for clusters
DO NOT USE WHEN REBASING CHROMIUM OS.

This simplifies swap block allocation for ZRAM.  Previously
the allocator tried to cluster blocks for both rotating and
solid-state disks.  Because of some performance bug, under
heavy swap usage scan_swap_map could take a large amount of
CPU time.  ZRAM has no use for clusters, so we use a plain
linear scan instead.  Since we never fill swap completely,
this is more than sufficient.

This performance bug exists at least on 3.4, 3.8, and 3.14.
I verified that this behavior is fixed in 3.19.  Not sure
about 3.18 but I think that the relevant swapfile.c fixes
are in there.

With this change, the fraction of samples on some heavy loads
(frequent tab switches with lots of tabs) goes from 25% of
the total to 0.2%.

I also ran a synthetic load on snow/lucas which walks 2.5GB
of memory in a random pattern.  The number of scan_swap_map()
samples goes from 7% to 0.

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

BUG=chromium:464949
TEST=see above

Change-Id: Id81b7083092731aa3d98daf903748f45aecf54eb
Reviewed-on: https://chromium-review.googlesource.com/262654
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>