CHROMIUM: mwifiex: scan less channels per command when connected
Our intention is to reduce number of channels scanned per scan
command when the device is connected. This will help us give
priority to Tx traffic and start serving it(if any) as soon
as currently downloaded scan command is received.

Current logic to decide "max_chan_per_scan" was designed assuming
the restriction of number of command nodes as 20. As we have
increased it to 50, we can have "max_chan_per_scan" as 1 always when
device is connected.

BUG=chrome-os-partner:38012
TEST=With all patches in this series applied,
verify network_WiFi_OverlappingBSSScan testcase passed on daisy.
Run complete wifi_matfunc suite.

Change-Id: I753f905bd24256087920192d137b6ec7cc6eb714
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Signed-off-by: Avinash Patil <patila@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/262849
Reviewed-by: Paul Stewart <pstew@chromium.org>