CHROMIUM: mwifiex: fix corner case connection failure
It is observed that when user issues "iw connect" command and
driver can not find requested SSID in scan results, all further
connection requests are rejected by cfg80211 with error
-EALREADY.

This is a regression issue introduced by below change which fixed
a smatch compilation error.
https://chromium-review.googlesource.com/#/c/58999/

This patch updates return variable when BSS is NULL to fix
the problem.

BUG=chrome-os-partner:37807
TEST=Confirm that when SSID is not found scan results, further
connection requests are not rejected by cfg80211.

Change-Id: I2f9ebea8c72aa363c79e40896fc5390a35fee968
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Signed-off-by: Avinash Patil <patila@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/264746
Reviewed-by: Paul Stewart <pstew@chromium.org>