CHROMIUM: ath9k: provide regulatory_hint() for world-sku chips
Because ath9k_btcoex registers with wireless core with
WIPHY_FLAG_STRICT_REGULATORY, wireless core will initially
not report regulatory domain changes to ath9k_btcoex. Instead,
wireless core will wait until at9k_btcoex reports a
regulatory_hint() to wireless core.

This works fine for non-world-sku chips, as a country code is
derived from EEPROM information, and reported as a regulatory_hint(),
in ath9k_init_device(). For world-sku chips, however, the call
to regulatory_hint() is skipped.

Add a call to regulatory_hint() for world chips, so that we can
inform the ath9k_btcoex driver if the system's VPD indicates
that the system is designated for a "constrained" regdomain. (See
regulatory-domain.conf for further details.)

BUG=chromium:480628
TEST=network_WiFi_RegDomain

Note: the new version of network_WiFi_RegDomain, which was used
to test this patch, will be uploaded shortly.

Change-Id: I73ab5ff64bfd20b600272a52da41e0c74cac96d6
Signed-off-by: quiche@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/267029
Reviewed-by: Paul Stewart <pstew@chromium.org>