CHROMIUM: improve hung task detection output
Commit d70d82212880ad380d2293e994d496bab8bd3258 added a "half way" hung
task warning, that is triggered when uninterruptible task was stuck for
60 seconds. Unfortunately the code was using WARN() and so the backtrace
that we are getting is backtrace of khungtaskd thread and not backtrace
of the hung task.

This change makes sure that we collect the right backtrace. We emulate
output of regular WARN() instance so that our crash collector (actually
warning collector) will pick it up and file it properly.

BUG=chromium:481227
TEST=Loaded Daisy so it produced hung task warning.

Change-Id: I7d2549bda7b288773f5efc5d4efbddfae9b7aa5c
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/269226
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>