CHROMIUM: mwifiex: do not get deadlocked on command timeout
If sending a command to the card times out right in the middle of
processing transmit request our work will try to claim the host and dump
the registers before trying to reset the card, but the host is already
claimed by the thread doing transmit. If firmware is truly stuck and
transmit never completes we'll be stuck until kernel hung task watchdog
kicks in and forcibly panics the box.

To work around the issue "try" claiming the host and eventually give up.
That will allow us to proceed with resetting the card which should kick
the transmit thread and unwedge the box.

BUG=chromium:488268
TEST=Used Daisy to stream YouTube videos over wireless.

Change-Id: Ifdb47b1874e7415f434173abea1abd800fbdd7c1
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/271306
Reviewed-by: Douglas Anderson <dianders@chromium.org>