devfreq: exynos542x: Bump up the LV_5 and LV_6 clocks of jpeg to 200MHz
With the jpeg standalone tests we have seen issues with
higher latency. This is due to devfreq not able to bump
up to higher freuqencies as the load is not perceived high
enough by the PPMU load monitor.

So we bump up the LV_5 and LV_6 level clocks of jpeg to 200MHz.

BUG=chrome-os-partner:40139
TEST=Run the jpeg standalone test and see average latency as 7.9ms.

Change-Id: I53062b916a02a05ab23f32901a273cd50cd1ba39
Signed-off-by: Arjun K V <arjun.kv@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/271463
Reviewed-by: Heng-ruey Hsu <henryhsu@google.com>
Tested-by: Heng-ruey Hsu <henryhsu@google.com>
Commit-Queue: Heng-ruey Hsu <henryhsu@google.com>