UPSTREAM: cfg80211: properly send NL80211_ATTR_DISCONNECTED_BY_AP in disconnect
When we disconnect from the AP, drivers call cfg80211_disconnect().
This doesn't know whether the disconnection was initiated locally
or by the AP though, which can cause problems with the supplicant,
for example with WPS. This issue obviously doesn't show up with any
mac80211 based driver since mac80211 doesn't call this function.

Fix this by requiring drivers to indicate whether the disconnect is
locally generated or not. I've tried to update the drivers, but may
not have gotten the values correct, and some drivers may currently
not be able to report correct values. In case of doubt I left it at
false, which is the current behaviour.

For libertas, make adjustments as indicated by Dan Williams.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 80279fb7ba5b71981a60988b0307afa43f78f6b1)

BUG=chrome-os-partner:48225
TEST=autotest network_WiFi_DisconnectReason

Change-Id: I4a93e86dfa3a50f76042f2f4682043d95a895bc6
Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/355821
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>